+++ arch/arm/src/imxrt/imxrt_flexcan.c
@@ -54,9 +54,7 @@
 
 #include <arch/board/board.h>
 
-#ifdef CONFIG_NET_CMSG
 #include <sys/time.h>
-#endif
 
 #ifdef CONFIG_IMXRT_FLEXCAN
 
@@ -93,11 +91,7 @@
 
 #define POOL_SIZE                   1
 
-#ifdef CONFIG_NET_CMSG
 #define MSG_DATA                    sizeof(struct timeval)
-#else
-#define MSG_DATA                    0
-#endif
 
 /* CAN bit timing values  */
 #define CLK_FREQ                    80000000
