+++ net/can/can_sockif.c
@@ -64,15 +64,6 @@ static int  can_accept(FAR struct socket *psock, FAR struct sockaddr *addr,
               FAR socklen_t *addrlen, FAR struct socket *newsock);
 static int  can_poll_local(FAR struct socket *psock, FAR struct pollfd *fds,
               bool setup);
-static ssize_t can_send(FAR struct socket *psock,
-              FAR const void *buf, size_t len, int flags);
-static ssize_t can_sendto(FAR struct socket *psock, FAR const void *buf,
-              size_t len, int flags, FAR const struct sockaddr *to,
-              socklen_t tolen);
-#ifdef CONFIG_NET_CMSG
-static ssize_t can_sendmsg(FAR struct socket *psock, FAR struct msghdr *msg,
-                    int flags);
-#endif
 static int can_close(FAR struct socket *psock);
 
 /****************************************************************************
@@ -91,16 +82,8 @@ const struct sock_intf_s g_can_sockif =
   can_connect,      /* si_connect */
   can_accept,       /* si_accept */
   can_poll_local,   /* si_poll */
-  can_send,         /* si_send */
-  can_sendto,       /* si_sendto */
-#ifdef CONFIG_NET_SENDFILE
-  NULL,             /* si_sendfile */
-#endif
-  can_recvfrom,     /* si_recvfrom */
-#ifdef CONFIG_NET_CMSG
-  can_recvmsg,      /* si_recvmsg */
   can_sendmsg,      /* si_sendmsg */
-#endif
+  can_recvmsg,      /* si_recvmsg */
   can_close         /* si_close */
 };
 
@@ -674,128 +657,6 @@ errout_with_lock:
 
   return ret;
 }
-
-/****************************************************************************
- * Name: can_send
- *
- * Description:
- *   The can_send() call may be used only when the socket is in
- *   a connected state  (so that the intended recipient is known).
- *
- * Input Parameters:
- *   psock - An instance of the internal socket structure.
- *   buf   - Data to send
- *   len   - Length of data to send
- *   flags - Send flags (ignored)
- *
- * Returned Value:
- *   On success, returns the number of characters sent.  On  error, a negated
- *   errno value is returned (see send() for the list of appropriate error
- *   values.
- *
- ****************************************************************************/
-
-static ssize_t can_send(FAR struct socket *psock, FAR const void *buf,
-                        size_t len, int flags)
-{
-  ssize_t ret;
-
-  /* Only SOCK_RAW is supported */
-
-  if (psock->s_type == SOCK_RAW)
-    {
-      /* Raw packet send */
-
-      ret = psock_can_send(psock, buf, len);
-    }
-  else
-    {
-      /* EDESTADDRREQ.  Signifies that the socket is not connection-mode and
-       * no peer address is set.
-       */
-
-      ret = -EDESTADDRREQ;
-    }
-
-  return ret;
-}
-
-/****************************************************************************
- * Name: can_sendto
- *
- * Description:
- *   If sendto() is used on a connection-mode (SOCK_STREAM, SOCK_SEQPACKET)
- *   socket, the parameters to and 'tolen' are ignored (and the error EISCONN
- *   may be returned when they are not NULL and 0), and the error ENOTCONN is
- *   returned when the socket was not actually connected.
- *
- * Input Parameters:
- *   psock    A reference to the socket structure of the socket
- *            to be connected
- *   buf      Data to send
- *   len      Length of data to send
- *   flags    Send flags (ignored)
- *   to       Address of recipient
- *   tolen    The length of the address structure
- *
- * Returned Value:
- *   None
- *
- * Assumptions:
- *
- ****************************************************************************/
-
-static ssize_t can_sendto(FAR struct socket *psock, FAR const void *buf,
-                          size_t len, int flags,
-                          FAR const struct sockaddr *to, socklen_t tolen)
-{
-  nerr("ERROR: sendto() not supported for raw packet sockets\n");
-  return -EAFNOSUPPORT;
-}
-
-/****************************************************************************
- * Name: can_sendmsg
- *
- * Description:
- *   The can_sendmsg() send a CAN frame to psock
- *
- * Input Parameters:
- *   psock - An instance of the internal socket structure.
- *   msg   - CAN frame and optional CMSG
- *   flags - Send flags (ignored)
- *
- * Returned Value:
- *   On success, returns the number of characters sent.  On  error, a negated
- *   errno value is returned (see send() for the list of appropriate error
- *   values.
- *
- ****************************************************************************/
-#ifdef CONFIG_NET_CMSG
-static ssize_t can_sendmsg(FAR struct socket *psock, FAR struct msghdr *msg,
-                    int flags)
-{
-  ssize_t ret;
-
-  /* Only SOCK_RAW is supported */
-
-  if (psock->s_type == SOCK_RAW)
-    {
-      /* Raw packet send */
-
-      ret = psock_can_sendmsg(psock, msg);
-    }
-  else
-    {
-      /* EDESTADDRREQ.  Signifies that the socket is not connection-mode and
-       * no peer address is set.
-       */
-
-      ret = -EDESTADDRREQ;
-    }
-
-  return ret;
-}
-#endif
 
 /****************************************************************************
  * Name: can_close
