+++ net/icmpv6/icmpv6_sockif.c
@@ -61,8 +61,6 @@ static int        icmpv6_accept(FAR struct socket *psock,
                     FAR struct socket *newsock);
 static int        icmpv6_netpoll(FAR struct socket *psock,
                     FAR struct pollfd *fds, bool setup);
-static ssize_t    icmpv6_send(FAR struct socket *psock, FAR const void *buf,
-                    size_t len, int flags);
 static int        icmpv6_close(FAR struct socket *psock);
 
 /****************************************************************************
@@ -81,16 +79,8 @@ const struct sock_intf_s g_icmpv6_sockif =
   icmpv6_connect,     /* si_connect */
   icmpv6_accept,      /* si_accept */
   icmpv6_netpoll,     /* si_poll */
-  icmpv6_send,        /* si_send */
-  icmpv6_sendto,      /* si_sendto */
-#ifdef CONFIG_NET_SENDFILE
-  NULL,               /* si_sendfile */
-#endif
-  icmpv6_recvfrom,    /* si_recvfrom */
-#ifdef CONFIG_NET_CMSG
-  NULL,               /* si_recvmsg */
-  NULL,               /* si_sendmsg */
-#endif
+  icmpv6_sendmsg,     /* si_sendmsg */
+  icmpv6_recvmsg,     /* si_recvmsg */
   icmpv6_close        /* si_close */
 };
 
@@ -450,35 +440,6 @@ static int icmpv6_netpoll(FAR struct socket *psock, FAR struct pollfd *fds,
 
       return icmpv6_pollteardown(psock, fds);
     }
-}
-
-/****************************************************************************
- * Name: icmpv6_send
- *
- * Description:
- *   Socket send() method for the raw packet socket.
- *
- * Input Parameters:
- *   psock    An instance of the internal socket structure.
- *   buf      Data to send
- *   len      Length of data to send
- *   flags    Send flags
- *
- * Returned Value:
- *   On success, returns the number of characters sent.  On  error, a negated
- *   errno value is returned (see send() for the list of appropriate error
- *   values.
- *
- ****************************************************************************/
-
-static ssize_t icmpv6_send(FAR struct socket *psock, FAR const void *buf,
-                        size_t len, int flags)
-{
-  /* ICMPv6 sockets cannot be bound and, hence, cannot support any
-   * connection-oriented data transfer.
-   */
-
-  return -EDESTADDRREQ;
 }
 
 /****************************************************************************
