+++ net/ieee802154/ieee802154_recvmsg.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * net/socket/ieee802154_recvfrom.c
+ * net/socket/ieee802154_recvmsg.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -272,43 +272,43 @@ static uint16_t
  ****************************************************************************/
 
 /****************************************************************************
- * Name: ieee802154_recvfrom
+ * Name: ieee802154_recvmsg
  *
  * Description:
  *   Implements the socket recvfrom interface for the case of the AF_INET
- *   and AF_INET6 address families.  ieee802154_recvfrom() receives messages
+ *   and AF_INET6 address families.  ieee802154_recvmsg() receives messages
  *   from a socket, and may be used to receive data on a socket whether or
  *   not it is connection-oriented.
  *
- *   If 'from' is not NULL, and the underlying protocol provides the source
- *   address, this source address is filled in.  The argument 'fromlen' is
- *   initialized to the size of the buffer associated with from, and
+ *   If msg_name is not NULL, and the underlying protocol provides the source
+ *   address, this source address is filled in. The argument 'msg_namelen' is
+ *   initialized to the size of the buffer associated with msg_name, and
  *   modified on return to indicate the actual size of the address stored
  *   there.
  *
  * Input Parameters:
  *   psock    A pointer to a NuttX-specific, internal socket structure
- *   buf      Buffer to receive data
- *   len      Length of buffer
+ *   msg      Buffer to receive the message
  *   flags    Receive flags
- *   from     Address of source (may be NULL)
- *   fromlen  The length of the address structure
  *
  * Returned Value:
- *   On success, returns the number of characters received.  If no data is
+ *   On success, returns the number of characters received. If no data is
  *   available to be received and the peer has performed an orderly shutdown,
- *   recv() will return 0.  Otherwise, on errors, a negated errno value is
- *   returned (see recvfrom() for the list of appropriate error values).
+ *   recvmsg() will return 0.  Otherwise, on errors, a negated errno value is
+ *   returned (see recvmsg() for the list of appropriate error values).
  *
  * Assumptions:
  *   The network is locked.
  *
  ****************************************************************************/
 
-ssize_t ieee802154_recvfrom(FAR struct socket *psock, FAR void *buf,
-                            size_t len, int flags, FAR struct sockaddr *from,
-                            FAR socklen_t *fromlen)
+ssize_t ieee802154_recvmsg(FAR struct socket *psock, FAR struct msghdr *msg,
+                           int flags)
 {
+  FAR void *buf = msg->msg_iov->iov_base;
+  size_t len = msg->msg_iov->iov_len;
+  FAR struct sockaddr *from = msg->msg_name;
+  FAR socklen_t *fromlen = &msg->msg_namelen;
   FAR struct ieee802154_conn_s *conn =
     (FAR struct ieee802154_conn_s *)psock->s_conn;
   FAR struct radio_driver_s *radio;
