+++ net/ieee802154/ieee802154_sockif.c
@@ -63,11 +63,6 @@ static int        ieee802154_accept(FAR struct socket *psock,
                     FAR struct socket *newsock);
 static int        ieee802154_poll_local(FAR struct socket *psock,
                     FAR struct pollfd *fds, bool setup);
-static ssize_t    ieee802154_send(FAR struct socket *psock,
-                   FAR const void *buf, size_t len, int flags);
-static ssize_t    ieee802154_sendto(FAR struct socket *psock,
-                   FAR const void *buf, size_t len, int flags,
-                   FAR const struct sockaddr *to, socklen_t tolen);
 static int        ieee802154_close(FAR struct socket *psock);
 
 /****************************************************************************
@@ -86,16 +81,8 @@ const struct sock_intf_s g_ieee802154_sockif =
   ieee802154_connect,     /* si_connect */
   ieee802154_accept,      /* si_accept */
   ieee802154_poll_local,  /* si_poll */
-  ieee802154_send,        /* si_send */
-  ieee802154_sendto,      /* si_sendto */
-#ifdef CONFIG_NET_SENDFILE
-  NULL,                   /* si_sendfile */
-#endif
-  ieee802154_recvfrom,    /* si_recvfrom */
-#ifdef CONFIG_NET_CMSG
-  NULL,                   /* si_recvmsg */
-  NULL,                   /* si_sendmsg */
-#endif
+  ieee802154_sendmsg,     /* si_sendmsg */
+  ieee802154_recvmsg,     /* si_recvmsg */
   ieee802154_close        /* si_close */
 };
 
@@ -615,123 +602,6 @@ static int ieee802154_poll_local(FAR struct socket *psock,
 
 #warning Missing logic
   return -ENOSYS;
-}
-
-/****************************************************************************
- * Name: ieee802154_send
- *
- * Description:
- *   Socket send() method for the PF_IEEE802154 socket.
- *
- * Input Parameters:
- *   psock    An instance of the internal socket structure.
- *   buf      Data to send
- *   len      Length of data to send
- *   flags    Send flags
- *
- * Returned Value:
- *   On success, returns the number of characters sent.  On  error, a negated
- *   errno value is returned (see send() for the list of appropriate error
- *   values.
- *
- ****************************************************************************/
-
-static ssize_t ieee802154_send(FAR struct socket *psock, FAR const void *buf,
-                               size_t len, int flags)
-{
-  struct sockaddr_ieee802154_s to;
-  FAR struct ieee802154_conn_s *conn;
-  ssize_t ret;
-
-  DEBUGASSERT(psock != NULL || buf != NULL);
-  conn = (FAR struct ieee802154_conn_s *)psock->s_conn;
-  DEBUGASSERT(conn != NULL);
-
-  /* Only SOCK_DGRAM is supported (because the MAC header is stripped) */
-
-  if (psock->s_type == SOCK_DGRAM)
-    {
-      /* send() may be used only if the socket has been connected. */
-
-      if (!_SS_ISCONNECTED(psock->s_flags) ||
-          conn->raddr.s_mode == IEEE802154_ADDRMODE_NONE)
-        {
-          ret = -ENOTCONN;
-        }
-      else
-        {
-          to.sa_family = AF_IEEE802154;
-          memcpy(&to.sa_addr, &conn->raddr,
-                 sizeof(struct ieee802154_saddr_s));
-
-          /* Then perform the send() as sendto() */
-
-          ret = psock_ieee802154_sendto(psock, buf, len, flags,
-                                        (FAR const struct sockaddr *)&to,
-                                        sizeof(
-                                          struct sockaddr_ieee802154_s));
-        }
-    }
-  else
-    {
-      /* EDESTADDRREQ.  Signifies that the socket is not connection-mode and
-       * no peer address is set.
-       */
-
-      ret = -EDESTADDRREQ;
-    }
-
-  return ret;
-}
-
-/****************************************************************************
- * Name: ieee802154_sendto
- *
- * Description:
- *   Implements the sendto() operation for the case of the PF_IEEE802154
- *   socket.
- *
- * Input Parameters:
- *   psock    A pointer to a NuttX-specific, internal socket structure
- *   buf      Data to send
- *   len      Length of data to send
- *   flags    Send flags
- *   to       Address of recipient
- *   tolen    The length of the address structure
- *
- * Returned Value:
- *   On success, returns the number of characters sent.  On  error, a negated
- *   errno value is returned (see send_to() for the list of appropriate error
- *   values.
- *
- ****************************************************************************/
-
-static ssize_t ieee802154_sendto(FAR struct socket *psock,
-                                 FAR const void *buf,
-                                 size_t len, int flags,
-                                 FAR const struct sockaddr *to,
-                                 socklen_t tolen)
-{
-  ssize_t ret;
-
-  /* Only SOCK_DGRAM is supported (because the MAC header is stripped) */
-
-  if (psock->s_type == SOCK_DGRAM)
-    {
-      /* Raw packet send */
-
-      ret = psock_ieee802154_sendto(psock, buf, len, flags, to, tolen);
-    }
-  else
-    {
-      /* EDESTADDRREQ.  Signifies that the socket is not connection-mode and
-       * no peer address is set.
-       */
-
-      ret = -EDESTADDRREQ;
-    }
-
-  return ret;
 }
 
 /****************************************************************************
