+++ net/local/Make.defs
@@ -23,15 +23,11 @@
 ifeq ($(CONFIG_NET_LOCAL),y)
 
 NET_CSRCS += local_conn.c local_release.c local_bind.c local_fifo.c
-NET_CSRCS += local_recvfrom.c local_sendpacket.c local_recvutils.c
-NET_CSRCS += local_sockif.c local_netpoll.c
+NET_CSRCS += local_recvmsg.c local_sendpacket.c local_recvutils.c
+NET_CSRCS += local_sockif.c local_netpoll.c local_sendmsg.c
 
 ifeq ($(CONFIG_NET_LOCAL_STREAM),y)
-NET_CSRCS += local_connect.c local_listen.c local_accept.c local_send.c
-endif
-
-ifeq ($(CONFIG_NET_LOCAL_DGRAM),y)
-NET_CSRCS += local_sendto.c
+NET_CSRCS += local_connect.c local_listen.c local_accept.c
 endif
 
 # Include Unix domain socket build support
