+++ net/pkt/pkt.h
@@ -199,39 +199,35 @@ uint16_t pkt_callback(FAR struct net_driver_s *dev,
 /* pkt_input() is prototyped in include/nuttx/net/pkt.h */
 
 /****************************************************************************
- * Name: pkt_recvfrom
+ * Name: pkt_recvmsg
  *
  * Description:
- *   Implements the socket recvfrom interface for the case of the AF_INET
- *   and AF_INET6 address families.  pkt_recvfrom() receives messages from
+ *   Implements the socket recvmsg interface for the case of the AF_INET
+ *   and AF_INET6 address families.  pkt_recvmsg() receives messages from
  *   a socket, and may be used to receive data on a socket whether or not it
  *   is connection-oriented.
  *
- *   If 'from' is not NULL, and the underlying protocol provides the source
- *   address, this source address is filled in.  The argument 'fromlen' is
- *   initialized to the size of the buffer associated with from, and
- *   modified on return to indicate the actual size of the address stored
- *   there.
+ *   If 'msg_name' is not NULL, and the underlying protocol provides the
+ *   source address, this source address is filled in.  The argument
+ *   'msg_namelen' is initialized to the size of the buffer associated with
+ *   msg_name, and modified on return to indicate the actual size of the
+ *   address stored there.
  *
  * Input Parameters:
  *   psock    A pointer to a NuttX-specific, internal socket structure
- *   buf      Buffer to receive data
- *   len      Length of buffer
+ *   msg      Buffer to receive the message
  *   flags    Receive flags
- *   from     Address of source (may be NULL)
- *   fromlen  The length of the address structure
  *
  * Returned Value:
- *   On success, returns the number of characters received.  If no data is
+ *   On success, returns the number of characters received. If no data is
  *   available to be received and the peer has performed an orderly shutdown,
- *   recv() will return 0.  Otherwise, on errors, a negated errno value is
- *   returned (see recvfrom() for the list of appropriate error values).
+ *   recvmsg() will return 0.  Otherwise, on errors, a negated errno value is
+ *   returned (see recvmsg() for the list of appropriate error values).
  *
  ****************************************************************************/
 
-ssize_t pkt_recvfrom(FAR struct socket *psock, FAR void *buf, size_t len,
-                     int flags, FAR struct sockaddr *from,
-                     FAR socklen_t *fromlen);
+ssize_t pkt_recvmsg(FAR struct socket *psock, FAR struct msghdr *msg,
+                    int flags);
 
 /****************************************************************************
  * Name: pkt_find_device
@@ -271,26 +267,26 @@ FAR struct net_driver_s *pkt_find_device(FAR struct pkt_conn_s *conn);
 void pkt_poll(FAR struct net_driver_s *dev, FAR struct pkt_conn_s *conn);
 
 /****************************************************************************
- * Name: psock_pkt_send
+ * Name: pkt_sendmsg
  *
  * Description:
- *   The psock_pkt_send() call may be used only when the packet socket is in
+ *   The pkt_sendmsg() call may be used only when the packet socket is in
  *   a connected state (so that the intended recipient is known).
  *
  * Input Parameters:
  *   psock    An instance of the internal socket structure.
- *   buf      Data to send
- *   len      Length of data to send
+ *   msg      Message to send
+ *   flags    Send flags
  *
  * Returned Value:
- *   On success, returns the number of characters sent.  On  error,
- *   a negated errno value is retruend.  See send() for the complete list
- *   of return values.
+ *   On success, returns the number of characters sent. On error, a negated
+ *   errno value is returned (see sendmsg() for the complete list of return
+ *   values.
  *
  ****************************************************************************/
 
-ssize_t psock_pkt_send(FAR struct socket *psock, FAR const void *buf,
-                       size_t len);
+ssize_t pkt_sendmsg(FAR struct socket *psock, FAR struct msghdr *msg,
+                    int flags);
 
 #undef EXTERN
 #ifdef __cplusplus
