+++ net/socket/Make.defs
@@ -23,6 +23,7 @@
 SOCK_CSRCS += bind.c connect.c getsockname.c getpeername.c
 SOCK_CSRCS += recv.c recvfrom.c send.c sendto.c
 SOCK_CSRCS += socket.c net_close.c
+SOCK_CSRCS += recvmsg.c sendmsg.c
 SOCK_CSRCS += net_dup2.c net_sockif.c net_poll.c net_vfcntl.c
 SOCK_CSRCS += net_fstat.c
 
@@ -54,8 +55,3 @@ endif
 DEPPATH += --dep-path socket
 VPATH += :socket
 
-# Support for control messages (CMSG)
-ifeq ($(CONFIG_NET_CMSG),y)
-SOCK_CSRCS += recvmsg.c
-SOCK_CSRCS += sendmsg.c
-endif
