+++ net/usrsock/usrsock_recvmsg.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * net/usrsock/usrsock_sendto.c
+ * net/usrsock/usrsock_recvmsg.c
  *
  *  Copyright (C) 2015, 2017 Haltian Ltd. All rights reserved.
  *  Author: Jussi Kivilinna <jussi.kivilinna@haltian.com>
@@ -58,105 +58,116 @@
  * Private Functions
  ****************************************************************************/
 
-static uint16_t sendto_event(FAR struct net_driver_s *dev, FAR void *pvconn,
-                             FAR void *pvpriv, uint16_t flags)
+static uint16_t recvfrom_event(FAR struct net_driver_s *dev,
+                               FAR void *pvconn, FAR void *pvpriv,
+                               uint16_t flags)
 {
-  FAR struct usrsock_reqstate_s *pstate = pvpriv;
+  FAR struct usrsock_data_reqstate_s *pstate = pvpriv;
   FAR struct usrsock_conn_s *conn = pvconn;
 
   if (flags & USRSOCK_EVENT_ABORT)
     {
       ninfo("socket aborted.\n");
 
-      pstate->result = -ECONNABORTED;
+      pstate->reqstate.result = -ECONNABORTED;
+      pstate->valuelen = 0;
+      pstate->valuelen_nontrunc = 0;
 
       /* Stop further callbacks */
 
-      pstate->cb->flags   = 0;
-      pstate->cb->priv    = NULL;
-      pstate->cb->event   = NULL;
+      pstate->reqstate.cb->flags   = 0;
+      pstate->reqstate.cb->priv    = NULL;
+      pstate->reqstate.cb->event   = NULL;
 
       /* Wake up the waiting thread */
 
-      nxsem_post(&pstate->recvsem);
+      nxsem_post(&pstate->reqstate.recvsem);
     }
   else if (flags & USRSOCK_EVENT_REQ_COMPLETE)
     {
       ninfo("request completed.\n");
 
-      pstate->result = conn->resp.result;
+      pstate->reqstate.result = conn->resp.result;
+      if (pstate->reqstate.result < 0)
+        {
+          pstate->valuelen = 0;
+          pstate->valuelen_nontrunc = 0;
+        }
+      else
+        {
+          pstate->valuelen = conn->resp.valuelen;
+          pstate->valuelen_nontrunc = conn->resp.valuelen_nontrunc;
+        }
 
-      if (pstate->result >= 0 || pstate->result == -EAGAIN)
+      if (pstate->reqstate.result >= 0 ||
+          pstate->reqstate.result == -EAGAIN)
         {
           /* After reception of data, mark input not ready. Daemon will
            * send event to restore this flag.
            */
 
-          conn->flags &= ~USRSOCK_EVENT_SENDTO_READY;
+          conn->flags &= ~USRSOCK_EVENT_RECVFROM_AVAIL;
         }
 
       /* Stop further callbacks */
 
-      pstate->cb->flags   = 0;
-      pstate->cb->priv    = NULL;
-      pstate->cb->event   = NULL;
+      pstate->reqstate.cb->flags   = 0;
+      pstate->reqstate.cb->priv    = NULL;
+      pstate->reqstate.cb->event   = NULL;
 
       /* Wake up the waiting thread */
 
-      nxsem_post(&pstate->recvsem);
+      nxsem_post(&pstate->reqstate.recvsem);
     }
   else if (flags & USRSOCK_EVENT_REMOTE_CLOSED)
     {
       ninfo("remote closed.\n");
 
-      pstate->result = -EPIPE;
+      pstate->reqstate.result = -EPIPE;
 
       /* Stop further callbacks */
 
-      pstate->cb->flags   = 0;
-      pstate->cb->priv    = NULL;
-      pstate->cb->event   = NULL;
+      pstate->reqstate.cb->flags   = 0;
+      pstate->reqstate.cb->priv    = NULL;
+      pstate->reqstate.cb->event   = NULL;
 
       /* Wake up the waiting thread */
 
-      nxsem_post(&pstate->recvsem);
+      nxsem_post(&pstate->reqstate.recvsem);
     }
-  else if (flags & USRSOCK_EVENT_SENDTO_READY)
+  else if (flags & USRSOCK_EVENT_RECVFROM_AVAIL)
     {
-      ninfo("sendto ready.\n");
+      ninfo("recvfrom avail.\n");
 
-      /* Do not let other waiters to claim new data. */
-
-      flags &= ~USRSOCK_EVENT_SENDTO_READY;
+      flags &= ~USRSOCK_EVENT_RECVFROM_AVAIL;
 
       /* Stop further callbacks */
 
-      pstate->cb->flags   = 0;
-      pstate->cb->priv    = NULL;
-      pstate->cb->event   = NULL;
+      pstate->reqstate.cb->flags   = 0;
+      pstate->reqstate.cb->priv    = NULL;
+      pstate->reqstate.cb->event   = NULL;
 
       /* Wake up the waiting thread */
 
-      nxsem_post(&pstate->recvsem);
+      nxsem_post(&pstate->reqstate.recvsem);
     }
 
   return flags;
 }
 
 /****************************************************************************
- * Name: do_sendto_request
+ * Name: do_recvfrom_request
  ****************************************************************************/
 
-static int do_sendto_request(FAR struct usrsock_conn_s *conn,
-                             FAR const void *buf, size_t buflen,
-                             FAR const struct sockaddr *addr,
-                             socklen_t addrlen, int32_t flags)
+static int do_recvfrom_request(FAR struct usrsock_conn_s *conn,
+                               size_t buflen, socklen_t addrlen,
+                               int32_t flags)
 {
-  struct usrsock_request_sendto_s req =
+  struct usrsock_request_recvfrom_s req =
   {
   };
 
-  struct iovec bufs[3];
+  struct iovec bufs[1];
 
   if (addrlen > UINT16_MAX)
     {
@@ -170,61 +181,71 @@ static int do_sendto_request(FAR struct usrsock_conn_s *conn,
 
   /* Prepare request for daemon to read. */
 
-  req.head.reqid = USRSOCK_REQUEST_SENDTO;
+  req.head.reqid = USRSOCK_REQUEST_RECVFROM;
   req.usockid = conn->usockid;
   req.flags = flags;
-  req.addrlen = addrlen;
-  req.buflen = buflen;
+  req.max_addrlen = addrlen;
+  req.max_buflen = buflen;
 
   bufs[0].iov_base = (FAR void *)&req;
   bufs[0].iov_len = sizeof(req);
-  bufs[1].iov_base = (FAR void *)addr;
-  bufs[1].iov_len = addrlen;
-  bufs[2].iov_base = (FAR void *)buf;
-  bufs[2].iov_len = buflen;
 
   return usrsockdev_do_request(conn, bufs, ARRAY_SIZE(bufs));
 }
 
 /****************************************************************************
- * Public Functions
- ****************************************************************************/
-
-/****************************************************************************
- * Name: usrsock_sendto
+ * Name: usrsock_recvmsg
  *
  * Description:
- *   If sendto() is used on a connection-mode (SOCK_STREAM, SOCK_SEQPACKET)
- *   socket, the parameters to and 'tolen' are ignored (and the error EISCONN
- *   may be returned when they are not NULL and 0), and the error ENOTCONN is
- *   returned when the socket was not actually connected.
+ *   recvmsg() receives messages from a socket, and may be used to receive
+ *   data on a socket whether or not it is connection-oriented.
+ *
+ *   If from is not NULL, and the underlying protocol provides the source
+ *   address, this source address is filled in. The argument fromlen
+ *   initialized to the size of the buffer associated with from, and modified
+ *   on return to indicate the actual size of the address stored there.
  *
  * Input Parameters:
- *   psock    A reference to the socket structure of the socket
- *   buf      Data to send
- *   len      Length of data to send
- *   flags    Send flags (ignored)
- *   to       Address of recipient
- *   tolen    The length of the address structure
+ *   psock    A pointer to a NuttX-specific, internal socket structure
+ *   msg      Buffer to receive the message
+ *   flags    Receive flags (ignored)
  *
  * Returned Value:
- *   On success, returns the number of characters sent.  On any failure, a
- *   negated errno value is returned.
+ *   On success, returns the number of characters received.  If no data is
+ *   available to be received and the peer has performed an orderly shutdown,
+ *   recvfrom() will return 0.  Otherwise, on any failure, a negated errno
+ *   value is returned.
  *
  ****************************************************************************/
 
-ssize_t usrsock_sendto(FAR struct socket *psock, FAR const void *buf,
-                       size_t len, int flags, FAR const struct sockaddr *to,
-                       socklen_t tolen)
+ssize_t usrsock_recvmsg(FAR struct socket *psock, FAR struct msghdr *msg,
+                        int flags)
 {
+  FAR void *buf = msg->msg_iov->iov_base;
+  size_t len = msg->msg_iov->iov_len;
+  FAR struct sockaddr *from = msg->msg_name;
+  FAR socklen_t *fromlen = &msg->msg_namelen;
   FAR struct usrsock_conn_s *conn = psock->s_conn;
-  struct usrsock_reqstate_s state =
+  struct usrsock_data_reqstate_s state =
   {
   };
 
+  struct iovec inbufs[2];
+  socklen_t addrlen = 0;
+  socklen_t outaddrlen = 0;
   ssize_t ret;
 
   DEBUGASSERT(conn);
+
+  if (fromlen)
+    {
+      if (*fromlen > 0 && from == NULL)
+        {
+          return -EINVAL;
+        }
+
+      addrlen = *fromlen;
+    }
 
   net_lock();
 
@@ -233,7 +254,7 @@ ssize_t usrsock_sendto(FAR struct socket *psock, FAR const void *buf,
     {
       /* Invalid state or closed by daemon. */
 
-      ninfo("usockid=%d; sendto() with uninitialized usrsock.\n",
+      ninfo("usockid=%d; recvfrom() with uninitialized usrsock.\n",
             conn->usockid);
 
       ret = (conn->state == USRSOCK_CONN_STATE_ABORTED) ? -EPIPE :
@@ -259,17 +280,12 @@ ssize_t usrsock_sendto(FAR struct socket *psock, FAR const void *buf,
             {
               /* Not connected. */
 
+              ninfo("usockid=%d; socket not connected.\n",
+                    conn->usockid);
+
               ret = -ENOTCONN;
               goto errout_unlock;
             }
-        }
-
-      if (to || tolen)
-        {
-          /* Address provided for connection-mode socket */
-
-          ret = -EISCONN;
-          goto errout_unlock;
         }
     }
 
@@ -277,7 +293,8 @@ ssize_t usrsock_sendto(FAR struct socket *psock, FAR const void *buf,
     {
       /* Non-blocking connecting. */
 
-      ninfo("usockid=%d; socket still connecting.\n", conn->usockid);
+      ninfo("usockid=%d; socket still connecting.\n",
+            conn->usockid);
 
       ret = -EAGAIN;
       goto errout_unlock;
@@ -289,51 +306,51 @@ ssize_t usrsock_sendto(FAR struct socket *psock, FAR const void *buf,
 
       if (conn->flags & USRSOCK_EVENT_REMOTE_CLOSED)
         {
-          ninfo("usockid=%d; remote closed.\n", conn->usockid);
+          ninfo("usockid=%d; remote closed (EOF).\n", conn->usockid);
 
-          ret = -EPIPE;
+          ret = 0;
           goto errout_unlock;
         }
 
-      /* Check if need to wait for send to become ready. */
+      /* Check if need to wait for receive data to become available. */
 
-      if (!(conn->flags & USRSOCK_EVENT_SENDTO_READY))
+      if (!(conn->flags & USRSOCK_EVENT_RECVFROM_AVAIL))
         {
           if (_SS_ISNONBLOCK(psock->s_flags) || (flags & MSG_DONTWAIT) != 0)
             {
-              /* Send busy at daemon side. */
+              /* Nothing to receive from daemon side. */
 
               ret = -EAGAIN;
               goto errout_unlock;
             }
 
-          /* Wait send to become ready. */
+          /* Wait recv to become avail. */
 
-          ret = usrsock_setup_request_callback(conn, &state, sendto_event,
-                                               USRSOCK_EVENT_ABORT |
-                                               USRSOCK_EVENT_SENDTO_READY |
-                                               USRSOCK_EVENT_REMOTE_CLOSED);
+          ret = usrsock_setup_data_request_callback(
+              conn, &state, recvfrom_event,
+              USRSOCK_EVENT_ABORT | USRSOCK_EVENT_RECVFROM_AVAIL |
+              USRSOCK_EVENT_REMOTE_CLOSED);
           if (ret < 0)
             {
               nwarn("usrsock_setup_request_callback failed: %d\n", ret);
               goto errout_unlock;
             }
 
-          /* Wait for send-ready (or abort, or timeout, or signal). */
+          /* Wait for receive-avail (or abort, or timeout, or signal). */
 
-          ret = net_timedwait(&state.recvsem,
-                              _SO_TIMEOUT(psock->s_sndtimeo));
+          ret = net_timedwait(&state.reqstate.recvsem,
+                              _SO_TIMEOUT(psock->s_rcvtimeo));
           if (ret < 0)
             {
               if (ret == -ETIMEDOUT)
                 {
-                  ninfo("sendto timedout\n");
+                  ninfo("recvfrom timedout\n");
 
                   ret = -EAGAIN;
                 }
               else if (ret == -EINTR)
                 {
-                  ninfo("sendto interrupted\n");
+                  ninfo("recvfrom interrupted\n");
                 }
               else
                 {
@@ -342,7 +359,7 @@ ssize_t usrsock_sendto(FAR struct socket *psock, FAR const void *buf,
                 }
             }
 
-          usrsock_teardown_request_callback(&state);
+          usrsock_teardown_data_request_callback(&state);
 
           /* Did wait timeout or got signal? */
 
@@ -363,47 +380,81 @@ ssize_t usrsock_sendto(FAR struct socket *psock, FAR const void *buf,
 
           if (conn->flags & USRSOCK_EVENT_REMOTE_CLOSED)
             {
-              ret = -EPIPE;
+              ret = 0;
               goto errout_unlock;
             }
 
-          DEBUGASSERT(conn->flags & USRSOCK_EVENT_SENDTO_READY);
+          DEBUGASSERT(conn->flags & USRSOCK_EVENT_RECVFROM_AVAIL);
         }
 
       /* Set up event callback for usrsock. */
 
-      ret = usrsock_setup_request_callback(conn, &state, sendto_event,
-                                           USRSOCK_EVENT_ABORT |
-                                           USRSOCK_EVENT_REQ_COMPLETE);
+      ret = usrsock_setup_data_request_callback(
+          conn, &state, recvfrom_event,
+          USRSOCK_EVENT_ABORT | USRSOCK_EVENT_REQ_COMPLETE);
       if (ret < 0)
         {
           nwarn("usrsock_setup_request_callback failed: %d\n", ret);
           goto errout_unlock;
         }
 
+      inbufs[0].iov_base = (FAR void *)from;
+      inbufs[0].iov_len = addrlen;
+      inbufs[1].iov_base = (FAR void *)buf;
+      inbufs[1].iov_len = len;
+
+      usrsock_setup_datain(conn, inbufs, ARRAY_SIZE(inbufs));
+
       /* MSG_DONTWAIT is only use in usrsock. */
 
-       flags &= ~MSG_DONTWAIT;
+      flags &= ~MSG_DONTWAIT;
 
       /* Request user-space daemon to close socket. */
 
-      ret = do_sendto_request(conn, buf, len, to, tolen, flags);
+      ret = do_recvfrom_request(conn, len, addrlen, flags);
       if (ret >= 0)
         {
           /* Wait for completion of request. */
 
-          net_lockedwait_uninterruptible(&state.recvsem);
-          ret = state.result;
+          net_lockedwait_uninterruptible(&state.reqstate.recvsem);
+          ret = state.reqstate.result;
 
           DEBUGASSERT(ret <= (ssize_t)len);
+          DEBUGASSERT(state.valuelen <= addrlen);
+          DEBUGASSERT(state.valuelen <= state.valuelen_nontrunc);
+
+          if (ret >= 0)
+            {
+              /* Store length of 'from' address that was available at
+               * daemon-side.
+               */
+
+              outaddrlen = state.valuelen_nontrunc;
+
+              /* If the MSG_PEEK flag is enabled, it will only peek
+               * from the buffer, so remark the input as ready.
+               */
+
+              if (flags & MSG_PEEK)
+                {
+                  conn->flags |= USRSOCK_EVENT_RECVFROM_AVAIL;
+                }
+            }
         }
 
-      usrsock_teardown_request_callback(&state);
+      usrsock_teardown_datain(conn);
+      usrsock_teardown_data_request_callback(&state);
     }
   while (ret == -EAGAIN);
 
 errout_unlock:
   net_unlock();
+
+  if (fromlen)
+    {
+      *fromlen = outaddrlen;
+    }
+
   return ret;
 }
